chore: re-structure monorepo shared packages
