--- @app/src/routes/+page.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-	export let data;
-	const queries = data.queries();
-</script>
-
-<pre>
-{JSON.stringify($queries, null, 2)}
-</pre>
-
