--- @eslint-shared/eslint.config.d.mts
@@ -1,3 +1,0 @@
-declare const _default: import('eslint').Linter.FlatConfig[];
-export default _default;
-//# sourceMappingURL=eslint.config.d.mts.map
