--- @eslint-shared/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "eslint-shared",
-  "version": "1.0.0",
-  "description": "Shared ESLint config for this monorepo",
-  "type": "module",
-  "main": "eslint.config.mjs",
-	"typings": "eslint.config.d.mts",
-  "devDependencies": {
-    "eslint-config-prettier": "^9.1.0",
-    "globals": "^15.0.0"
-  }
-}
