+++ @example/eslint.config.js
@@ -1,6 +1,6 @@
 import ts from 'typescript-eslint';
 import svelte from 'eslint-plugin-svelte';
-import shared from 'eslint-shared';
+import shared from 'shared/eslint';
 
 /** @type {import('eslint').Linter.FlatConfig[]} */
 export default [
