+++ @example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "app",
+  "name": "example",
   "version": "0.0.1",
   "private": true,
   "description": "SvelteKit app for testing trpc-svelte-query-adapter",
