+++ @example/src/app.d.ts
