+++ @example/src/lib/index.ts
