+++ @example/src/routes/+layout.svelte
