+++ @example/src/routes/+layout.ts
