+++ @example/src/routes/+page.ts
@@ -1,0 +1,16 @@
+import { trpc } from '$lib/trpc/client';
+
+export async function load(event) {
+	const { queryClient } = await event.parent();
+	const api = trpc(event, queryClient);
+
+	return {
+		query: await api.greeting.createServerQuery('foo'),
+		// queries: await api.createServerQueries((t) => [t.greeting('foo'), t.greeting('bar')], {
+		// 	combine: (results) => ({
+		// 		data: results.map(({ data }) => data),
+		// 		pending: results.some(({ isPending }) => isPending),
+		// 	}),
+		// }),
+	};
+}
