+++ @example/static/favicon.png
