+++ @example/svelte.config.js
