+++ @example/vite.config.ts
