--- @prettier-shared/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "prettier-shared",
-  "version": "1.0.0",
-  "description": "Shared Prettier config for this monorepo",
-  "type": "module",
-  "main": "prettier.config.mjs",
-  "typings": "prettier.config.d.mts",
-  "devDependencies": {
-    "prettier": "^3.1.1"
-  }
-}
