+++ @shared/eslint/index.d.ts
@@ -1,0 +1,3 @@
+declare const _default: import("eslint").Linter.FlatConfig[];
+export default _default;
+//# sourceMappingURL=index.d.ts.map
