+++ @shared/eslint/index.js
