+++ @shared/prettier/index.d.ts
@@ -1,3 +1,3 @@
 declare const _default: import("prettier").Config;
 export default _default;
-//# sourceMappingURL=prettier.config.d.mts.map
+//# sourceMappingURL=index.d.ts.map
