+++ package.json
@@ -10,7 +10,6 @@
     "typescript": "^5.0.0"
   },
   "dependencies": {
-    "eslint-shared": "workspace:^",
-    "prettier-shared": "workspace:^"
+    "shared": "workspace:^"
   }
 }
